backend/requirement.txt
amqp
apiclient
beautifulsoup4
billiard
boto3
botocore
bs4
cachetools
celery
certifi
chardet
click
colorama
defusedxml
Django
django-allauth
django-anymail
django-bootstrap
django-bootstrap3
django-celery-beat
django-celery-results
django-cors-headers
django-enumfield
django-environ
django-extensions
django-uni-form
djangorestframework
docutils
eventlet
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
greenlet
html5lib
httpie
httplib2
idna
jmespath
json-table-schema
kombu
lxml
messytables
mysqlclient
numpy
oauth2client
oauthlib
pandas
psycopg2
pyasn1
pyasn1-modules
Pygments
PyPDF2
python-dateutil
python-dotenv
python-magic
python-memcached
python3-openid
pytz
redis
requests
requests-oauthlib
rsa
s3transfer
see
selenium
six
SQLAlchemy
sqlalchemy-redshift
tabula-py
tinys3
uritemplate
urllib3
vine
webencodings
Werkzeug
xlrd
